clean up drivetrain, pose2d, vector2d, rotation2d

+++ TeamCode/src/main/java/org/ftc3825/GVF/Path.kt
@@ -57,7 +57,7 @@ class Path(vararg var pathSegments: PathSegment) {
             (
                 currentPath.endHeading
                 -(
-                    ( (heading + PI / 2) + PI ) % ( 2 * PI ) - PI
+                    ( (heading.toDouble() + PI / 2) + PI ) % ( 2 * PI ) - PI
                 )
                 + PI
             ) % ( 2 * PI ) - PI
